feat(phase-e2): MC 1.3-1.4 bracket — blight boss, witch/blight-skeleton/bat, beacon, ender chest, adventure mode, lava fluid, effects system
Version-evolution Phase 2 (Adventure Features). All values live-verified
2026-09-06 against the reference wiki (research record:
docs/research/phase2-1.3-1.4-research.md, ~120 VERIFIED citations).

- blight.rs: summon (T of spirit sand + 3 skulls), 220-tick charge, 300 HP,
  1 HP/20-t regen, 2 s skull cadence (8 HP + Blight II 10 s Normal / 40 s
  Hard), 3x4x3 block breaking on damage, hollow star + 50 XP drops
- effects.rs: timed Blight/Poison/Regeneration + beacon stat effects
  (speed/strength/resistance/jump multipliers per the w/Effect rows)
- beacon.rs: pyramid scan 1-4 (9/34/83/164 blocks), power gating,
  4 s reapply at 9+2xlevel s, range 20/30/40/50, feed cycle
- mobs: blight skeleton (fortress platform #2 spawner), witch (~0.97%
  monster pool), bat (ambient, light<=3 below sea level, groups of 8)
- ender chest: shared 27-slot container, breaks into 8 obsidian
- adventure mode (GameType 2): no block break/place, interactions open
- anvil: 3 damage stages (12%/use), gravity block; repair GUI deferred
  (no damageable items — disclosed)
- lava fluid: dimension-aware spread (OW 3/30-tick, Hollow 7/10-tick),
  fluid-quad meshing with the fixed lava tint, 4 HP/10-tick damage
- emerald ore: Mountains-only generation, drops 1 emerald + 3-7 XP
- foods: potato/baked/carrot/pumpkin pie (verified hunger values)
- u16 state-space widening (the <=255 window was exhausted) + latent E1
  World::set_block state-storage bug fixed
- render-distance slider 2-32 (VERIFICATION-REPORT fix)

Suite 342 -> 372 (+30), wasm32 clean.

+++ voxelcraft/crates/vc-anvil/src/save.rs
@@ -1243,7 +1243,7 @@ mod tests {
         let bytes = nbt::write_root("", &root).unwrap();
 
         let (chunk, _light) = chunk_from_nbt(&bytes).unwrap();
-        assert_eq!(chunk.get(0, 0, 0), STONE);
+        assert_eq!(chunk.get(0, 0, 0), STONE as u16);
         // chunk.get returns the raw u8 state id (57 = OAK_LOG_X); the flat
         // u16 accessor confirms the full state survived
         assert_eq!(chunk.get(1, 0, 0) as u16, OAK_LOG_X);
@@ -1439,7 +1439,7 @@ mod tests {
 
         // --- …the chunk reloads with the edit intact…
         let (loaded, _l) = load_chunk(&dir, 0, 0).unwrap().expect("chunk present");
-        assert_eq!(loaded.get(8, 70, 8), GLOWSTONE);
+        assert_eq!(loaded.get(8, 70, 8), GLOWSTONE as u16);
 
         // --- …and every other block matches a same-seed regeneration exactly
         let (fresh, _) = vc_world::gen::TerrainGen::new(seed).generate_chunk(0, 0, Vec::new());
